fixed routes, handlers and controllers
