+++ api/src/handlers/videogames.handlers.js
@@ -1,18 +1,18 @@
 const { createVideogame, searchGameByName, getGameById, getAllGames } = require ('../controllers/videogames.controllers.js');
 const {Videogame} = require('../db');
 
-const createVideogameHandler = async (req, res) => {
-  const {name, description, image, price, platform, genre} = req.body;
+// const createVideogameHandler = async (req, res) => {
+//   const {name, description, image, price, platform, genre} = req.body;
 
-  try {
-    const newGame = await createVideogame(name, description, image, price, platform, genre);
-    res.status(201).json(newGame);
-  } catch (error) {
-    res.status(400).json({error: error.message});
-  }
-};
+//   try {
+//     const newGame = await createVideogame(name, description, image, price, platform, genre);
+//     res.status(201).json(newGame);
+//   } catch (error) {
+//     res.status(400).json({error: error.message});
+//   }
+// };
 
-const videogamesHandlers = async (req, res) => {
+const getVideogames = async (req, res) => {
   const { name } = req.query;
   try {
   const results = name ? await searchGameByName(name) : await getAllGames();
@@ -22,7 +22,7 @@ const videogamesHandlers = async (req, res) => {
   }
 };
 
-const videogameHandler = async (req, res) => {
+const getVideogameById = async (req, res) => {
   const { id } = req.params;
   try {
       const game = await getGameById(id);
@@ -32,23 +32,6 @@ const videogameHandler = async (req, res) => {
   }
       
 };
-
-const getGameById = async (value) => {
-  if(value.length>5){
-      const findDbID= await Videogame.findByPk(value, {include:Genre, Platform})
-      const detailOfGame = {
-          id:findDbID.id,
-          name: findDbID.name,
-          description: findDbID.description,
-          image: findDbID.img,
-          price: findDbID.price,
-          platform: findDbID.platform.map(m=>m.name),
-          genre: findDbID.genre.map(m=>m.name)
-      }
-      return detailOfGame
-  }
-};
-
 
 const modifyVideogameHandler = async (req, res) => {
   const{ name, description, price } = req.body;
@@ -65,9 +48,8 @@ const modifyVideogameHandler = async (req, res) => {
   }
 };
 
-module.exports = {
-  createVideogameHandler, 
-  videogamesHandlers, 
-  videogameHandler, 
-  modifyVideogameHandler
+module.exports = { 
+  getVideogames,  
+  modifyVideogameHandler,
+  getVideogameById
 };
