+++ api/src/routes/videogames.js
@@ -1,8 +1,7 @@
 const Router = require('express');
-const { createVideogameDb, createVideogame, getVideogames, getVideogameById } = require('../controllers/videogames.controllers');
+const { getVideogames, getVideogameById } = require('../handlers/videogames.handlers');
+const { createVideogameDb, createVideogame } = require('../controllers/videogames.controllers');
 const router = Router();
-const {Videogame, Genre, Platform} = require("../db");
-const {videogames} = require('../utils/data-videogames');
 
 router.get('/', getVideogames);
 router.get('/:id', getVideogameById);
