+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "bcrypt": "^5.1.0"
+  }
+}
